+++ packages/github-action/action-dist/index.js
@@ -51367,7 +51367,7 @@ function renderSarifReport(result) {
                 tool: {
                     driver: {
                         name: "ProofPR",
-                        informationUri: "https://github.com/linsk27/proof-pr",
+                        informationUri: "https://github.com/linsk-labs/proof-pr",
                         rules: [...rules.values()]
                     }
                 },